Make getHumanChoice allow for more valid inputs
Previously getHumanChoice could only be answerd by "R", "P", "S" or
those lowercase. Now humanChoice is lowecased to lowerChoice.
The switch in the function now uses lowerChoice as the expression and
the cases are changed to have the first letter and full word be
valid case-insensitive inputs.
Also fixed some typos in the comments