Add playRound function and pseudocode for later
The playRound function takes humanChoice and computerChoice as
parameters, which will be the human and computers selection.
The paramaters are then compared through an if else statement to see
who wins. Adds one to the winners score and then returns who won as a
string. Also added pseudocode for the playGame function that playRound
will be declared in.