Add playGame function and loop inside
The playGame function declares the score variables and the playRound
function. It has a for loop in it that loops variable round, that starts
at one, and increases by one each loop until round equals sever.
In the loop if round is less than or equal to 5 it will first check
if round is between two and six and if so state the round in the console
and reassign the selection variables again, break. Then call playRound,
and state the current score. Else If round is equal to 6, initialize
the winner variable and assign the winner through a ternary. If winner
equals tie, log to console tie game and if not state the winner.