Add changeScore function and edit playRound
The changeScore function will be called in the playRound function once
a winner has been determined in the round.
The function takes in the result parameter, which determines who won the
round. If human won, then add one to humanScore else add one to
computerScore.

The function will also determine if a player has won the game.
When changeScore is called, after adding to the score, an if statement
checks if humanScore or computerScore is at 5 or above. If it is, then
return changing score to say who won in the DOM. If a score is not at 5
or above then the function will display the current scoreline on score.